Firebase deepLink add and card renewal purchased order and create purchase api integration

+++ app/src/main/java/com/nowfloats/PreSignUp/SplashScreen_Activity.java
@@ -4,18 +4,13 @@ import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-import android.provider.Settings;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.ImageView;
 
 import com.airbnb.lottie.LottieAnimationView;
 import com.airbnb.lottie.LottieDrawable;
-import com.boost.presignup.SignUpActivity;
 import com.boost.presignup.utils.PresignupManager;
-import com.bumptech.glide.Glide;
+import com.boost.upgrades.UpgradeActivity;
 import com.nowfloats.Login.Fetch_Home_Data;
 import com.nowfloats.Login.Login_MainActivity;
 import com.nowfloats.Login.Model.FloatsMessageModel;
@@ -24,7 +19,6 @@ import com.nowfloats.NavigationDrawer.API.GetVisitorsAndSubscribersCountAsyncTas
 import com.nowfloats.NavigationDrawer.HomeActivity;
 import com.nowfloats.signup.UI.Model.Get_FP_Details_Event;
 import com.nowfloats.signup.UI.Service.Get_FP_Details_Service;
-import com.nowfloats.signup.UI.UI.PreSignUpActivity;
 import com.nowfloats.sync.DbController;
 import com.nowfloats.test.com.nowfloatsui.buisness.util.Util;
 import com.nowfloats.util.BusProvider;
@@ -43,22 +37,27 @@ import java.util.ArrayList;
 import retrofit.RetrofitError;
 import retrofit.client.Response;
 
-public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.Fetch_Home_Data_Interface , PresignupManager.SignUpLoginHandler {
+public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.Fetch_Home_Data_Interface, PresignupManager.SignUpLoginHandler {
     UserSessionManager session;
     Bus bus;
     LottieAnimationView animationView;
     public static ProgressDialog pd;
     private String loginCheck = null, deepLink;
+    private String deepLinkViewType = "", deepLinkFpId = "";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash_screen_);
         Methods.isOnline(SplashScreen_Activity.this);
-
+        Bundle bundle = getIntent().getExtras();
         if (getIntent() != null && getIntent().getStringExtra("from") != null) {
             MixPanelController.track(EventKeysWL.NOTIFICATION_CLICKED, null);
             deepLink = getIntent().getStringExtra("url");
+        }
+        if (bundle != null) {
+            deepLinkViewType = bundle.getString("deepLinkViewType");
+            deepLinkFpId = bundle.getString("deepLinkFpId");
         }
         bus = BusProvider.getInstance().getBus();
         session = new UserSessionManager(getApplicationContext(), SplashScreen_Activity.this);
@@ -67,8 +66,8 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
 
     private Thread mThread;
 
-    private void initLottieAnimation(){
-        if(animationView == null)
+    private void initLottieAnimation() {
+        if (animationView == null)
             animationView = findViewById(R.id.pre_dashboard_animation);
         animationView.setAnimation(R.raw.pre_dashboard_lottie);
         animationView.setRepeatCount(LottieDrawable.INFINITE);
@@ -122,12 +121,7 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
 
     private void fetchData() {
         try {
-//            new Thread(new Runnable() {
-//                @Override
-//                public void run() {
-                    Util.addBackgroundImages();
-//                }
-//            }).start();
+            Util.addBackgroundImages();
             getFPDetails_retrofit(SplashScreen_Activity.this, session.getFPID(), Constants.clientId, bus);
         } catch (Exception e) {
             e.printStackTrace();
@@ -179,7 +173,7 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        if(animationView != null)
+        if (animationView != null)
             animationView.cancelAnimation();
         animationView = null;
         mThread = null;
@@ -187,51 +181,69 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
 
     @Subscribe
     public void post_getFPDetails(Get_FP_Details_Event response) {
-
 //        API_Business_enquiries businessEnquiries = new API_Business_enquiries(null,session);
 //        businessEnquiries.getMessages();
-
-        //VISITOR and SUBSCRIBER COUNT API
+//        VISITOR and SUBSCRIBER COUNT API
         if (Methods.isOnline(this)) {
             GetVisitorsAndSubscribersCountAsyncTask visit_subcribersCountAsyncTask = new GetVisitorsAndSubscribersCountAsyncTask(SplashScreen_Activity.this, session);
             visit_subcribersCountAsyncTask.execute();
         }
+        if (!deepLinkViewType.isEmpty() && deepLinkViewType.equalsIgnoreCase("CART_FRAGMENT")) {
+            initiateAddonMarketplace();
+        } else {
+            Intent i = new Intent(SplashScreen_Activity.this, HomeActivity.class);
+            i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
 
-
-        Intent i = new Intent(SplashScreen_Activity.this, HomeActivity.class);
-        i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-
-        if(deepLink!=null){
-            if(!deepLink.contains("logout")) {
-                //i.putExtra("url", deepLink);
-                i.putExtras(getIntent());
+            if (deepLink != null) {
+                if (!deepLink.contains("logout")) {
+                    i.putExtras(getIntent());
+                    startActivity(i);
+                    if (pd != null && pd.isShowing()) pd.dismiss();
+                    finish();
+                } else {
+                    session.logoutUser();
+                    DataBase db = new DataBase(this);
+                    DbController.getDbController(getApplicationContext()).deleteDataBase();
+                    db.deleteLoginStatus();
+                }
+            } else {
                 startActivity(i);
-                if (pd != null && pd.isShowing())
-                    pd.dismiss();
+                if (pd != null && pd.isShowing()) pd.dismiss();
                 finish();
-            }else {
-                session.logoutUser();
-                DataBase db = new DataBase(this);
-                DbController.getDbController(getApplicationContext()).deleteDataBase();
-                db.deleteLoginStatus();
             }
-        }else{
-            startActivity(i);
-            if (pd != null && pd.isShowing())
-                pd.dismiss();
-            finish();
         }
-        // Staring Login Activity
-
-
+        //Staring Login Activity
         /*Fetch_Home_Data fetch_home_data  = new Fetch_Home_Data(this,0);
         fetch_home_data.setFetchDataListener(SplashScreen_Activity.this);
         fetch_home_data.getMessages(session.getFPID(), "0");*/
     }
 
+    private void initiateAddonMarketplace() {
+        Intent intent = new Intent(SplashScreen_Activity.this, UpgradeActivity.class);
+        intent.putExtra("expCode", session.getFP_AppExperienceCode());
+        intent.putExtra("fpName", session.getFPName());
+        intent.putExtra("fpid", session.getFPID());
+        intent.putExtra("isFirebaseDeepLink", true);
+        intent.putExtra("deepLinkViewType", deepLinkViewType);
+        intent.putExtra("loginid", session.getUserProfileId());
+        if (session.getFPEmail() != null) {
+            intent.putExtra("email", session.getFPEmail());
+        } else {
+            intent.putExtra("email", "ria@nowfloats.com");
+        }
+        if (session.getFPPrimaryContactNumber() != null) {
+            intent.putExtra("mobileNo", session.getFPPrimaryContactNumber());
+        } else {
+            intent.putExtra("mobileNo", "9160004303");
+        }
+        intent.putExtra("profileUrl", session.getFPLogo());
+        startActivity(intent);
+        overridePendingTransition(0, 0);
+        finish();
+    }
+
     @Subscribe
-    public void getResponse(Response response)
-    {
+    public void getResponse(Response response) {
         Intent i = new Intent(SplashScreen_Activity.this, HomeActivity.class);
         i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
         if (pd != null && pd.isShowing())
@@ -241,8 +253,7 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
     }
 
     @Subscribe
-    public void getError(RetrofitError retrofitError)
-    {
+    public void getError(RetrofitError retrofitError) {
         Intent i = new Intent(SplashScreen_Activity.this, HomeActivity.class);
         i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
         if (pd != null && pd.isShowing())
@@ -250,6 +261,7 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
         startActivity(i);
         finish();
     }
+
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
@@ -278,21 +290,21 @@ public class SplashScreen_Activity extends Activity implements Fetch_Home_Data.F
         Intent i = new Intent(SplashScreen_Activity.this, HomeActivity.class);
         i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
         // Staring Login Activity
-        if(deepLink!=null){
-            if(!deepLink.contains("logout")) {
+        if (deepLink != null) {
+            if (!deepLink.contains("logout")) {
                 //i.putExtra("url", deepLink);
                 i.putExtras(getIntent());
                 startActivity(i);
                 if (pd != null)
                     pd.dismiss();
                 finish();
-            }else {
+            } else {
                 session.logoutUser();
                 DataBase db = new DataBase(this);
                 DbController.getDbController(getApplicationContext()).deleteDataBase();
                 db.deleteLoginStatus();
             }
-        }else{
+        } else {
             startActivity(i);
             if (pd != null)
                 pd.dismiss();
